+++ .claude.exs
@@ -1,10 +1,8 @@
 %{
   hooks: %{
-    session_start: [:deps_get],
     stop: [:compile, :format],
-    subagent_stop: [:compile, :format],
     post_tool_use: [:compile, :format],
-    # These are only ran on git commit
-    pre_tool_use: [:compile, :format, :unused_deps]
+    pre_tool_use: [:compile, :format, :unused_deps],
+    subagent_stop: [:compile, :format]
   }
 }